Improve main101 and rename some parameters
